refactor: extract comment card
