+++ app/etkinlik/[id].js
@@ -12,6 +12,7 @@ import axiosClient from '../../src/api/axiosClient';
 import { Linking } from 'react-native';
 import * as Clipboard from 'expo-clipboard';
 import { KeyboardAvoidingView } from 'react-native';
+import CommentCard from '../../src/components/CommentCard';
 
 const PRIMARY = '#7B2CBF';
 const ACCENT = '#FFD54F';
@@ -81,22 +82,9 @@ export default function EtkinlikDetay() {
           Alert.alert("Paylaşım başarısız", "Geçersiz paylaşım tipi");
           break;
       }
-    } catch (err) {
+  } catch (err) {
       Alert.alert("Hata", "Paylaşım gerçekleştirilemedi.");
     }
-  };
-
-  const zamanFarki = (tarih) => {
-    const simdi = new Date();
-    const yorumTarihi = new Date(tarih);
-    const saniye = Math.floor((simdi - yorumTarihi) / 1000);
-    if (saniye < 60) return `${saniye} sn önce`;
-    const dakika = Math.floor(saniye / 60);
-    if (dakika < 60) return `${dakika} dk önce`;
-    const saat = Math.floor(dakika / 60);
-    if (saat < 24) return `${saat} sa önce`;
-    const gun = Math.floor(saat / 24);
-    return `${gun} gün önce`;
   };
 
   const yanitGonder = async (yorumId) => {
@@ -118,157 +106,10 @@ export default function EtkinlikDetay() {
       setYanitlar(prev => ({ ...prev, [yorumId]: '' }));
       setYanitId(null);
     } catch (err) {
-      Alert.alert("Hata", "Yanıt gönderilemedi.");
+      Alert.alert('Hata', 'Yanıt gönderilemedi.');
     }
   };
 
-  const touchableIcon = (liked, onPress, count, tarih) => (
-    <View style={{ flexDirection: 'row', alignItems: 'center', gap: 6 }}>
-      <TouchableOpacity onPress={onPress}>
-        <Text style={{ fontSize: 18 }}>{liked ? '💜' : '🤍'}</Text>
-      </TouchableOpacity>
-      <Text style={{ fontSize: 12, color: '#888' }}>{count}</Text>
-      <Text style={{ fontSize: 12, color: '#aaa' }}>· {zamanFarki(tarih)}</Text>
-    </View>
-  );
-
-  const YorumCard = ({ yorum, level = 0 }) => {
-    const inputRef = useRef(null);
-
-useEffect(() => {
-  if (yanitId !== yorum._id) return;
-  const timeout = setTimeout(() => {
-    if (inputRef.current?.isFocused?.() === false) {
-      inputRef.current.focus();
-    }
-  }, 100);
-  return () => clearTimeout(timeout);
-}, [yanitId]);
-
-    const altYorumlar = getYanitlar(yorum._id);
-    const isAltYorum = level === 1;
-    const isUser = auth?.userId === yorum.kullaniciId?.toString();
-
-    const [duzenleModu, setDuzenleModu] = useState(false);
-    const [duzenlenenMetin, setDuzenlenenMetin] = useState(yorum.yorum);
-
-    const yorumuGuncelle = async () => {
-      try {
-        const { data } = await axiosClient.put(`/yorum/${yorum._id}`, {
-          yorum: duzenlenenMetin,
-        });
-        setYorumlar(prev => prev.map(y => (y._id === yorum._id ? data : y)));
-        setDuzenleModu(false);
-      } catch {
-        Alert.alert("Hata", "Yorum güncellenemedi");
-      }
-    };
-
-    const yorumuSil = async () => {
-      Alert.alert("Yorum silinsin mi?", "Bu işlem geri alınamaz", [
-        { text: "İptal", style: "cancel" },
-        {
-          text: "Sil", style: "destructive", onPress: async () => {
-            try {
-              await axiosClient.delete(`/yorum/${yorum._id}`);
-              setYorumlar(prev => prev.filter(y => y._id !== yorum._id));
-            } catch {
-              Alert.alert("Hata", "Silme başarısız");
-            }
-          },
-        },
-      ]);
-    };
-
-    return (
-      <View key={yorum._id} style={{ marginBottom: 12 }}>
-        <View style={{ flexDirection: 'row', paddingLeft: isAltYorum ? 48 : 0 }}>
-          <Image source={{ uri: yorum.avatarUrl }} style={{ width: 40, height: 40, borderRadius: 20, marginRight: 10 }} />
-          <View style={{ flex: 1 }}>
-            <Text style={{ fontWeight: '600' }}>{yorum.kullanici}</Text>
-
-            {duzenleModu ? (
-              <>
-                <TextInput
-                  value={duzenlenenMetin}
-                  onChangeText={setDuzenlenenMetin}
-                  style={{ borderWidth: 1, borderColor: '#ccc', borderRadius: 6, padding: 8, marginVertical: 4 }}
-                  multiline
-                />
-                <TouchableOpacity onPress={yorumuGuncelle} style={{ backgroundColor: PRIMARY, padding: 6, borderRadius: 6 }}>
-                  <Text style={{ color: '#fff' }}>Kaydet</Text>
-                </TouchableOpacity>
-              </>
-            ) : (
-              <Text style={{ color: '#555', marginVertical: 4 }}>{yorum.yorum}</Text>
-            )}
-
-            <View style={{ flexDirection: 'row', alignItems: 'center', justifyContent: 'space-between' }}>
-              {touchableIcon(yorum.begendinMi, async () => {
-                try {
-                  const { data: updated } = await axiosClient.put(`/yorum/begen/${yorum._id}`);
-                  setYorumlar(prev => prev.map(item => (item._id === updated._id ? updated : item)));
-                } catch {}
-              }, yorum.begeni || 0, yorum.tarih)}
-
-              <View style={{ flexDirection: 'row', gap: 8 }}>
-                <TouchableOpacity onPress={() => setYanitId(yorum._id)}>
-                  <Text style={{ color: PRIMARY }}>Yanıtla</Text>
-                </TouchableOpacity>
-              </View> 
-            </View>
-<View style={{ marginTop: 8, display: String(yanitId) === String(yorum._id) ? 'flex' : 'none' }}>
-  <TextInput
-    ref={inputRef}
-    value={yanitlar[yorum._id] || ''}
-    onChangeText={text => {
-      if (yanitlar[yorum._id] !== text) {
-        setYanitlar(prev => ({ ...prev, [yorum._id]: text }));
-      }
-    }}
-    placeholder="Yanıtınızı yazın..."
-    multiline
-    scrollEnabled
-    blurOnSubmit={false}
-    style={{
-      borderWidth: 1,
-      borderColor: '#ccc',
-      borderRadius: 8,
-      padding: 10,
-      minHeight: 60,
-    }}
-  />
-  <TouchableOpacity
-    onPress={() => yanitGonder(yorum._id)}
-    style={{
-      backgroundColor: PRIMARY,
-      borderRadius: 8,
-      paddingVertical: 10,
-      marginTop: 10,
-      alignSelf: 'flex-end',
-      width: 100,
-    }}
-  >
-    <Text style={{ color: '#fff', textAlign: 'center', fontWeight: '600' }}>
-      Yanıtla
-    </Text>
-  </TouchableOpacity>
-</View>
-
-
-
-
-
-
-
-          </View>
-        </View>
-        {altYorumlar.map(alt => (
-          <YorumCard key={alt._id} yorum={alt} level={1} />
-        ))}
-      </View>
-    );
-  };
 
   useEffect(() => {
     const yorumlariGetir = async () => {
@@ -591,7 +432,18 @@ const gorselSrc = etkinlik.gorsel?.startsWith('http') ? etkinlik.gorsel : `${bac
             <Text style={{ color: '#aaa' }}>Henüz yorum yapılmamış.</Text>
           ) : (
             yorumlar.filter(y => !y.ustYorumId).map(y => (
-              <YorumCard key={y._id} yorum={y} level={0} />
+              <CommentCard
+                key={y._id}
+                yorum={y}
+                level={0}
+                getYanitlar={getYanitlar}
+                yanitId={yanitId}
+                setYanitId={setYanitId}
+                yanitlar={yanitlar}
+                setYanitlar={setYanitlar}
+                yanitGonder={yanitGonder}
+                setYorumlar={setYorumlar}
+              />
             ))
           )}
